tv: hdmi: Fix build warning in hdmi driver
Fix the following build warning in exynos' hdmi driver:

drivers/media/video/exynos/tv/hdmi_drv.c:61:3: warning: initialization makes
pointer from integer without a cast [enabled by default]
drivers/media/video/exynos/tv/hdmi_drv.c:61:3: warning: (near initialization
for 'exynos_hdmi_match[0].data') [enabled by default]

BUG=chrome-os-partner:9696
TEST=build and boot on SMDK

Reviewed-on: https://gerrit-int.chromium.org/17825
Commit-Ready: Doug Anderson <dianders@google.com>
Tested-by: Doug Anderson <dianders@google.com>

+++ drivers/media/video/exynos/tv/hdmi_drv.c
@@ -58,7 +58,7 @@ static struct hdmi_driver_data hdmi_driver_data[] = {
 static const struct of_device_id exynos_hdmi_match[] = {
 	{
 		.compatible = "samsung,exynos5-hdmi",
-		.data = (unsigned long)&hdmi_driver_data[2],
+		.data = &hdmi_driver_data[2],
 	},
 	{},
 };
